add side length labels
