+++ packages/perseus/src/widgets/interactive-graphs/graphs/components/text-label.tsx
@@ -9,7 +9,7 @@ type Props = Pick<TextProps, "x" | "y" | "attach" | "attachDistance"> & {
 
 export const TextLabel = ({children, ...rest}: Props) => (
     <Text
-        size={15}
+        size={16}
         svgTextProps={{
             filter: "url(#background)",
             fontWeight: "bold",
@@ -18,4 +18,14 @@ export const TextLabel = ({children, ...rest}: Props) => (
     >
         {children}
     </Text>
+);
+
+export const SvgDefs = () => (
+    <defs>
+        {/* This filter is used to create a background for text labels */}
+        <filter id="background" x="-5%" width="110%" y="0%" height="100%">
+            <feFlood floodColor="#FFF" floodOpacity="0.64" />
+            <feComposite operator="over" in="SourceGraphic" />
+        </filter>
+    </defs>
 );
