+++ packages/perseus/src/widgets/interactive-graphs/mafs-graph.tsx
@@ -5,6 +5,7 @@ import * as React from "react";
 
 import GraphLockedLayer from "./graph-locked-layer";
 import {LinearGraph, PolygonGraph, RayGraph, SegmentGraph} from "./graphs";
+import {SvgDefs} from "./graphs/components/text-label";
 import {Grid} from "./grid";
 import {LegacyGrid} from "./legacy-grid";
 import {interactiveGraphReducer} from "./reducer/interactive-graph-reducer";
@@ -12,7 +13,6 @@ import {
     getGradableGraph,
     initializeGraphState,
 } from "./reducer/interactive-graph-state";
-import {SvgDefs} from "./svg-defs";
 
 import type {InteractiveGraphAction} from "./reducer/interactive-graph-action";
 import type {InteractiveGraphProps, InteractiveGraphState} from "./types";
